# app/microservice/public-swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GFW OGR API",
    "description": "Global Forest Watch - OGR API",
    "version": "1.0.0"
  },
  "host": "example.vizzuality.com",
  "schemes": [
    "https",
    "http"
  ],
  "produces": [
    "application/vnd.api+json"
  ],
  "paths": {
    "/ogr/convert": {
      "post": {
        "summary": "Convert from shapefile and others to geojson",
        "description": "Convert Shapefile to geojson",
        "produces": [
          "application/vnd.api+json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "tags": [
          "OGR"
        ],
        "operationId": "convert",
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "description": "File to convert",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "The object created",
            "schema": {
              "$ref": "#/definitions/GeoStoreAPI"
            }
          },
          "400": {
            "description": "Validate error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Errors": {
      "type": "object",
      "description": "Errors",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "A unique identifier for this particular occurrence of the problem."
        },
        "links": {
          "type": "object",
          "description": "A links object",
          "properties": {
            "about": {
              "type": "string",
              "description": "A link that leads to further details about this particular occurrence of the problem."
            }
          }
        },
        "status": {
          "type": "string",
          "description": "The HTTP status code applicable to this problem, expressed as a string value"
        },
        "code": {
          "type": "string",
          "description": "An application-specific error code, expressed as a string value"
        },
        "title": {
          "type": "string",
          "description": "A short, human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the problem, except for purposes of localization."
        },
        "detail": {
          "type": "string",
          "description": "A human-readable explanation specific to this occurrence of the problem. Like title, this field's value can be localized"
        },
        "source": {
          "type": "object",
          "description": "An object containing references to the source of the error, optionally including any of the following members",
          "properties": {
            "pointer": {
              "type": "string",
              "description": "A JSON Pointer [RFC6901] to the associated entity in the request document"
            },
            "parameter": {
              "type": "string",
              "description": "A string indicating which URI query parameter caused the error."
            }
          }
        },
        "meta": {
          "type": "object",
          "description": "A meta object containing non-standard meta-information about the error."
        }
      }
    },
    "GeoStoreAPI": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GeoJSON"
          }
        }
      }
    },
    "GeoJSON": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "description": "Resource type"
        },
        "id": {
          "type": "string",
          "description": "Resource id"
        },
        "attributes": {
          "properties": {
            "schema": {
              "$ref": "#/definitions/Features"
            }
          }
        }
      }
    },
    "Features": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "description": "Type of feature"
        },
        "features": {
          "type": "array",
          "description": "Features",
          "items": {
            "$ref": "#/definitions/Feature"
          }
        }
      }
    },
    "Feature": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "description": "Type of feature"
        },
        "properties": {
          "type": "object",
          "description": "Feature"
        },
        "geometry": {
          "type": "object",
          "description": "Geometries",
          "properties": {
            "schema": {
              "$ref": "#/definitions/Geometry"
            }
          }
        }
      }
    },
    "Geometry": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "description": "Geometry type"
        },
        "coordinates": {
          "type": "array",
          "description": "Coordinates array",
          "items": {
            "type": "object"
          }
        }
      }
    }
  }
}